[configure-splash-screen] feat: move utils from ./helpers to ./utils
