+++ packages/configure-splash-screen/src/utils/StateManager.ts
