+++ packages/configure-splash-screen/src/utils/file-utils.ts
