+++ packages/configure-splash-screen/src/utils/string-utils.ts
